Divided hexstody-hot into separate crates
