+++ hexstody-hot-operator/Cargo.toml
@@ -1,0 +1,57 @@
+[package]
+name = "hexstody-hot-operator"
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "operator-keygen"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+chrono = { version = "0.4.19", features = ["serde"] }
+clap = { version = "3.0.0-rc.4", features = ["derive", "env"] }
+env_logger = { version = "0.9.0" }
+futures = "0.3.19"
+futures-channel = "0.3"
+futures-util = "0.3.19"
+hexstody-db = { path = "../hexstody-db" }
+hexstody-api = { path = "../hexstody-api" }
+hexstody-btc-api = { path = "../hexstody-btc-api" }
+log = "0.4.14"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+sqlx = { version = "0.5", features = [
+  "runtime-tokio-rustls",
+  "migrate",
+  "macros",
+  "postgres",
+  "json",
+  "chrono",
+] }
+thiserror = "1.0"
+tokio = { version = "1", features = ["full"] }
+anyhow = "1.0"
+p256 = { version = "0.10.1", features = ["pem"] }
+pkcs8 = { version = "0.8.0", features = ["pem", "encryption", "std", "pkcs5"] }
+rand_core = { version = "0.6.3", features = ["std"] }
+rocket = { version = "0.5.0-rc.1", default-features = false, features = [ "json", "secrets" ] }
+rocket_okapi = { git = "https://github.com/GREsau/okapi", branch = "master", features = [ "swagger" ] }
+uuid = { version = "0.8.2", features = ["v4"] }
+rocket_dyn_templates = { version = "0.1.0-rc.1", features = ["handlebars"] }
+schemars = "0.8.8"
+pwhash = "1.0.0"
+rpassword = "6.0.1"
+ctrlc = "3.2.2"
+
+[dev-dependencies]
+bitcoin = "0.28.1"
+bitcoin_hashes = "0.10.0"
+bitcoincore-rpc = "0.15.0"
+bitcoincore-rpc-json = "0.15.0"
+maplit = "1.0.2"
+nix = "0.24.1"
+port-selector = "0.1.4"
+run_script = "0.9.0"
+sqlx-database-tester = { version = "0.2.0", features = ["runtime-tokio"] }
+tempdir = "0.3.7"
