+++ hexstody-hot-operator/Rocket.toml
@@ -1,0 +1,2 @@
+[default]
+template_dir = "templates"
