+++ hexstody-hot-operator/src/bin/operator-keygen.rs
