+++ hexstody-hot-operator/templates/index.html.hbs
@@ -1,11 +1,11 @@
 {{#*inline "css"}}
-<link rel="stylesheet" type="text/css" href="/css/operator/style.css" />
+<link rel="stylesheet" type="text/css" href="/css/style.css" />
 {{/inline}}
 
 {{#*inline "scripts"}}
-<script src="/scripts/operator/jscu.bundle.js"></script>
-<script src="/scripts/operator/jscec.bundle.js"></script>
-<script src="/scripts/operator/page/index.js"></script>
+<script src="/scripts/jscu.bundle.js"></script>
+<script src="/scripts/jscec.bundle.js"></script>
+<script src="/scripts/index.js"></script>
 {{/inline}}
 
 {{#*inline "page"}}
@@ -58,4 +58,4 @@
 </table>
 {{/inline}}
 
-{{> operator/base}}
+{{> base}}
