+++ hexstody-hot-public/Cargo.toml
@@ -1,10 +1,7 @@
 [package]
-name = "hexstody-hot"
+name = "hexstody-hot-public"
 version = "0.1.0"
 edition = "2021"
-
-[[bin]]
-name = "operator-keygen"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -34,14 +31,9 @@ sqlx = { version = "0.5", features = [
 ] }
 thiserror = "1.0"
 tokio = { version = "1", features = ["full"] }
-anyhow = "1.0"
-p256 = { version = "0.10.1", features = ["pem"] }
-pkcs8 = { version = "0.8.0", features = ["pem", "encryption", "std", "pkcs5"] }
-rand_core = { version = "0.6.3", features = ["std"] }
 rocket = { version = "0.5.0-rc.1", default-features = false, features = [ "json", "secrets" ] }
 rocket_okapi = { git = "https://github.com/GREsau/okapi", branch = "master", features = [ "swagger" ] }
 rocket_session = "0.2.2"
-uuid = { version = "0.8.2", features = ["v4"] }
 rocket_dyn_templates = { version = "0.1.0-rc.1", features = ["handlebars"] }
 schemars = "0.8.8"
 pwhash = "1.0.0"
