+++ hexstody-hot-public/Rocket.toml
@@ -1,0 +1,2 @@
+[default]
+template_dir = "templates"
