+++ hexstody-hot-public/src/api/public/auth.rs
