+++ hexstody-hot-public/static/scripts/page/common.js
