+++ hexstody-hot-public/static/scripts/page/signup.js
