+++ hexstody-hot-public/templates/footer.html.hbs
