--- hexstody-hot/src/api/mod.rs
@@ -1,2 +1,0 @@
-pub mod operator;
-pub mod public;
